+++ CHANGELOG.md
@@ -9,6 +9,11 @@ Note that since we don't clearly distinguish between a public and private interf
 - Add Charmm saccharide names
 - Treat missing occupancy column as occupany of 1
 - Fix line shader not accounting for aspect ratio
+- [Breaking] Fix point repr & shader
+    - Was unusable with ``wboit``
+    - Replaced ``pointFilledCircle`` & ``pointEdgeBleach`` params by ``pointStyle`` (square, circle, fuzzy)
+    - Set ``pointSizeAttenuation`` to false by default
+    - Set ``sizeTheme`` to ``uniform`` by default
 
 ## [v2.3.0] - 2021-09-06